convert one TI AIC script format to another
The PurePath DSP tool generates AIC scripts for loading the miniDSP,
but it uses the continuation character “>”, and the PurePath Wireless
Configurator doesn’t allow those in the AIC script format to set up
peripherals.  So this script will collect all the continuations and
make one big multibyte write with it.  Leaves all other lines untouched.